Package information on the site is now database-driven
The website is now data-driven.
Package information on the homepage stats slice is db driven.
The ten most-recently created packages in the db are linked to the homepage.
Package service is improved.
User service is created.
Load-data script in the project to load sample data from the real pypi. (This data is from 2018).
Database is modified.