Cookie Auth, server-side, validation, and restructuring.
The view-model structure is implemented. As a result, data is now handled through a viewmodel structure, simplifying the code of the files located in the views subdirectory (a.k.a. the view-controllers). Every viewmodel has a request_dict which serves to bring all the request arguments and form arguments into a single dictionary, which can then be rendered in the html templates. Server-side validation has been implemented within the viewmodels. (Client-side validation is next).
User sessions are authenticated by a cookie now. Passlib is now a requirement.